LayoutTests/imported/w3c/web-platform-tests/intersection-observer/observer-callback-arguments-expected.txt


FAIL Callback is invoked with |this| value of IntersectionObserver and two arguments assert_equals: expected (object) object "[object IntersectionObserver]" but got (undefined) undefined
